requirements.txt
asgiref==3.4.1
Django==3.2.18
django-filter==21.1
djangorestframework==3.14.0
importlib-metadata==4.8.3
Markdown==3.3.7
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.1.1
zipp==3.6.0